// src/app/models/project.ts
export class Project {
  email: string;
  projectName: string;
  projectOwner: string;
  customerName: string;
  companySite: string;
  phone: string;
  language: string;
  timeZone: string;
  addressLine1: string;
  addressLine2: string;
  postcode: string;
  city: string;
  state: string;
  country: string;
}